refactor EndPointForcesSinusoidal example and test cases
