+++ examples/JointCases/hinge_joint.py
@@ -6,9 +6,6 @@ import sys
 # FIXME without appending sys.path make it more generic
 sys.path.append("../../")
 from elastica import *
-from examples.JointCases.external_force_class_for_joint_test import (
-    EndpointForcesSinusoidal,
-)
 from examples.JointCases.joint_cases_postprocessing import (
     plot_position,
     plot_video,
